requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.7
PyQt5>=5.15.9
PyQtWebEngine>=5.15.6
requests>=2.28.2
colorama>=0.4.6
python-dotenv>=1.0.0
sse-starlette>=1.6.1
pyttsx3
gtts
playsound
pypiwin32
SpeechRecognition
langchain-openai
aiohttp
sentence-transformers
scikit-learn
numpy
beautifulsoup4
google-search-results
firebase-admin 